requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
black==21.10b0
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-celery-beat==2.2.1
django-redis==5.0.0
django-templated-mail==1.1.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.2
kombu==5.2.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
platformdirs==2.4.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
PyJWT==2.3.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
redis==3.5.3
regex==2021.11.2
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
